インターフェイスを分離
