+++ Assets/App/Scripts/Map/Chara/ICharaController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a6515390f66a4f2cab895838d7a48a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
